Extract util function to a seperate file
